Support both short and long form for secrets
